requirements.txt
flask
pytesseract
pillow